requirements.txt
langchain==0.3.9
langchain-community==0.3.1
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.128
streamlit==1.35.0
pandas==2.1.4
python-dotenv==1.0.0
openai==1.42.0
langchain==0.3.9
langchain-community==0.3.1
faiss-cpu==1.7.4
neo4j==5.25.0
tiktoken==0.7.0
transformers==4.41.2
sentence-transformers==3.0.1
requests==2.31.0